requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
pydantic==2.10.3
pyyaml==6.0.2
docker==7.1.0
pytz==2024.2
